Fix compile warnings: Woverloaded-virtual, real bugs, unused code
Stage 0/0b/1 of plans/warning-cleanup.md, bringing the clean-build
warning count from 564 to 495 (down to just -Wsign-conversion and
-Wsign-compare, tracked as stage 2/3).

- third_party/ is now a SYSTEM include path so vendored tinyxml2/
  pocketfft headers are exempt from our warning flags; tinyxml2.cpp
  (compiled directly, not just included) gets -w instead.
- Added several zero-violation -Werror= flags (duplicated-cond/
  duplicated-branches/logical-op/redundant-decls/cast-align/format)
  plus -Wshadow/-Wmissing-declarations (fixed their handful of real
  sites). -Wnull-dereference deliberately left off: false positives
  at -O2 through inlined STL container code, not a real bug class
  here - documented in a CMakeLists.txt comment.
- Command::values_[] could be read uninitialized on a malformed
  string_view (assert() is compiled out in Release); gave it a
  default member initializer.
- Fixed two real -Woverloaded-virtual sites by renaming rather than
  adding a using-declaration to suppress them: a differently-shaped
  render() in a derived class was hiding, not overriding, the base
  one. InstrumentTrackState::render(int) -> renderVoices(int)
  (ArpeggiatorState's override and all callers updated to match);
  SongState::render(int, Song&, Mixer&) -> renderBlock(...) (same,
  plus ~20 comment references across the codebase updated so none
  are left naming the old method).
- Went through every unused-variable/unused-but-set-variable site
  individually: most were dead pure-getter locals; PatternEditor's
  was_playing only fed an undocumented #if 0 auto-play-toggle block
  (removed together); main.cpp's --relative flag has done nothing
  since the first commit that introduced main.cpp and isn't part of
  the documented CLI surface (removed rather than left lying around).
- Fixed TerminalUI.cpp's ncinput narrowing casts, Chart.h's member
  init order, and Song.cpp's tinyxml2 SetAttribute sign-promo.

Verified: clean rebuild (0 errors), full ctest pass, and a --render
smoke test, since several changes touch the playback path.

+++ Song.cpp
@@ -25,7 +25,7 @@
 #include "bus/BusEffectRegistry.h"
 #include "MemoryParameterSource.h"
 
-#include "third_party/tinyxml2/tinyxml2.h"
+#include <tinyxml2/tinyxml2.h>
 
 #include "constants.h"
 
@@ -526,7 +526,7 @@ Song::save(const std::string & filename) const {
 	auto data = to_string(command);
 
 	auto command_element = doc.NewElement("command");
-	command_element->SetAttribute("row", row);
+	command_element->SetAttribute("row", static_cast<int>(row));
 	command_element->SetAttribute("data", data.c_str());
 	pattern_element->InsertEndChild(command_element);
       }
